requirements.txt
numpy>=2.0
Pillow>=10.0
matplotlib>=3.8
torch>=2.0
